requirements.txt
beautifulsoup4
fastapi
uvicorn
pillow
PyMuPDF
pytesseract

